# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.UpsertCommunityLinkRequest:
    properties:
      label:
        type: string
      platform:
        type: string
      url:
        type: string
    required:
    - platform
    - url
    type: object
...
  handler.setOverrideRequest:
    properties:
      limit_ngn:
        minimum: 1
        type: integer
    required:
    - limit_ngn
    type: object
...
  models.CommunityLink:
    properties:
      created_at:
        type: string
      id:
        type: string
      label:
        type: string
      platform:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  models.KYCTier:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - KYCTierNone
    - KYCTier1
    - KYCTier2
    - KYCTier3
...
  service.UserLimitInfo:
    properties:
      effective_limit_ngn:
        type: integer
      override_limit_ngn:
        description: null when no override
        type: integer
      remaining_ngn:
        type: integer
      tier:
        $ref: '#/definitions/models.KYCTier'
      tier_limit_ngn:
        type: integer
      today_usage_ngn:
        type: integer
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready backend API for Kynettic platform with authentication,
    user management, and health monitoring
  title: Kynettic API
  version: "1.0"
paths:
  /admin/community-links:
    post:
      consumes:
      - application/json
      description: 'Admin-only: add or update a community link (upsert by platform).'
      parameters:
      - description: Community link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpsertCommunityLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CommunityLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Upsert community link
      tags:
      - admin
...
  /admin/users/{id}/withdrawal-limit:
    get:
      description: 'Admin-only: returns tier limit, custom override, today''s usage,
        and remaining headroom.'
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserLimitInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get user withdrawal limit (admin)
      tags:
      - admin
  /admin/users/{id}/withdrawal-limit/override:
    delete:
      description: 'Admin-only: removes the custom daily withdrawal limit for a user,
        reverting to their KYC tier default.'
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Remove withdrawal limit override (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 'Admin-only: permanently sets a custom daily NGN withdrawal limit
        for a user, overriding their tier default.'
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New limit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.setOverrideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Set withdrawal limit override (admin)
      tags:
      - admin
...
  /community-links:
    get:
      description: Returns all community links. Requires authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CommunityLink'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: List community links
      tags:
      - community
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"